+++ plugins/modules/tfc_project_create.py
@@ -6,26 +6,28 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from __future__ import absolute_import, division, print_function
+from ansible.module_utils.basic import AnsibleModule
+from ..module_utils.tfc import TfcClient, TfcError, add_tags
 
 __metaclass__ = type
 
 
 DOCUMENTATION = '''
 ---
-module: tfc_workspace_create
+module: tfc_project_create
 
-short_description: Terraform Cloud API (HCP Terraform) module to create a workspace.
+short_description: Terraform Cloud API (HCP Terraform) module to create a project.
 
 version_added: 2.2.0
 
 description:
-  - This module create a workspace inside an organization
-  - See https://developer.hashicorp.com/terraform/cloud-docs/api-docs/workspaces#create-a-workspace
+  - This module create a project inside an organization
+  - See https://developer.hashicorp.com/terraform/cloud-docs/api-docs/projects#create-a-project
 
 options:
     organization:
         description:
-            - The name of the organization the workspace belongs to.
+            - The name of the organization the project belongs to.
         type: str
 
 extends_documentation_fragment:
@@ -38,31 +40,28 @@ author:
 '''
 
 EXAMPLES = '''
-- name: Create a workspace
-  tfc_workspace_create:
+- name: Create a project
+  tfc_project_create:
     token: "{{ lookup('ansible.builtin.env', 'TERRA_TOKEN') }}"
     payload:
       data:
         attributes:
-          type: "workspace"
-          name: "New_wk_name"
+          type: "project"
+          name: "New_project_name"
 '''
 
 RETURN = '''
 data:
     description:
-        - The data attribute from HCP Terraform route C(POST /organizations/:orga/workspaces)
+        - The data attribute from HCP Terraform route C(POST /organizations/:orga/projects)
     returned: success
     type: dict
 '''
 
-from ansible.module_utils.basic import AnsibleModule
-from ..module_utils.tfc import TfcClient, TfcError, add_tags
+PROJECT_PATH_BY_ORGANIZATION = "/organizations/{organization}/projects"
 
 
-WORKSPACE_PATH_BY_ORGANIZATION = "/organizations/{organization}/workspaces"
-
-def create_workspace(module_params):
+def create_project(module_params):
     organization = module_params.get('organization')
     validate_certs = module_params.get('validate_certs')
     token = module_params.get('api_token')
@@ -73,19 +72,21 @@ def create_workspace(module_params):
     payload = module_params.get('payload')
     tags = module_params.get('tags')
 
-    path = WORKSPACE_PATH_BY_ORGANIZATION.format(organization=organization)
+    path = PROJECT_PATH_BY_ORGANIZATION.format(organization=organization)
 
     if payload is not None:
         pass
     elif data is not None:
-        payload = {"data": {"type": "workspaces", **data}}
+        payload = {"data": {"type": "projects", **data}}
     else:
-        payload = {"data": {"type": "workspaces" }}
-
+        payload = { "data": {"type": "projects"} }
+    
     if attributes is not None:
         payload["data"]["attributes"] = attributes
     if tags is not None:
         add_tags(payload, tags)
+
+
 
     client = TfcClient(token, url=api_url)
     r = client.create(path, json=payload, verify=validate_certs,
@@ -96,7 +97,7 @@ def create_workspace(module_params):
 
 def main():
     """
-    Module tfc_workspace_create
+    Module tfc_project_create
     """
 
     argument_spec = dict(
@@ -120,7 +121,7 @@ def main():
     )
 
     try:
-        result = create_workspace(module.params)
+        result = create_project(module.params)
     except TfcError as e:
         module.fail_json(msg=str(e))